Главная страница
- хэндлер для главной страницы
- переводы вынесены в отдельный пакет
- шаблон для шапки страниц
- файл переводов для главной страницы
