+++ handlers/mainpage.go
@@ -4,6 +4,8 @@ import (
 	"log"
 	"net/http"
 	"text/template"
+
+	"github.com/dmidokov/remontti-v2/translations"
 )
 
 // Функция обработчик для главной страницы
@@ -15,7 +17,9 @@ func mainPage(w http.ResponseWriter, r *http.Request) {
 	files := []string{
 		rootPath + "main.page.gohtml",
 		rootPath + "base.layout.gohtml",
-		rootPath + "footer.partial.gohtml",
+		rootPath + "footers/footer.partial.gohtml",
+		rootPath + "headers/mainpage.partial.gohtml",
+		rootPath + "bodies/mainpage.partial.gohtml",
 	}
 
 	ts, err := template.ParseFiles(files...)
@@ -25,7 +29,16 @@ func mainPage(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	err = ts.Execute(w, nil)
+	pageData.Title = "Меню"
+	pageData.Translation, err = translations.GetTranslations("mainpage", cfg)
+
+	if err != nil {
+		log.Println(err.Error())
+		http.Error(w, "Internal Server Error", 500)
+		return
+	}
+
+	err = ts.Execute(w, pageData)
 	if err != nil {
 		log.Print(err)
 		http.Error(w, "Internal Server Error", 500)
