--- web/css/variables1.css
@@ -1,18 +1,0 @@
-:root {
-    --background : #d8d6d0;
-    --menu-background: #f1f3f5;
-    --modal-background: #f1efea;
-
-    --elem-1: #a8bec5;
-    --elem-2: #996c80;
-    --elem-3: #b69a7d;
-    --elem-4:#312f41;
-
-    --icon-1:#d48426;
-    --icon-2:#679b62;
-    --icon-3:#f44639;
-    --icon-4: #a1b9c0;
-    --icon-5: #949494;
-
-
-}
