+++ web/ui/bodies/mainpage.partial.gohtml
@@ -1,0 +1,5 @@
+{{define "main"}}
+<main>
+    body
+</main>
+{{end}}
