+++ web/ui/headers/mainpage.partial.gohtml
@@ -1,0 +1,22 @@
+{{define "header"}}
+<header class="main-header">
+    <div>{{ .Translation.PageTitle }}</div>
+    <div>---Навигация--</div>
+    <div class="left-side-menu">
+        <div class="dropdown">
+            <img class="header-icon" src="/static/images/account.png">
+            <ul class="dropdown-menu">
+                <li>
+                    <a href="#">{{ .Translation.Settings }}</a>
+                </li>
+                <li>
+                    <a href="/logout">{{ .Translation.Logout}}</a>
+                </li>
+            </ul>
+        </div>
+        <div>
+            <img class="header-icon" src="/static/images/settings.png">
+        </div>
+    </div>
+</header>
+{{end}}
