+++ web/ui/main.page.gohtml
@@ -1,7 +1,2 @@
 {{template "base" .}}
 
-{{define "main"}}
-<main>
-    Some data in main
-</main>
-{{end}}
