+++ web/ui/translations/ru_RU/mainpage
@@ -1,0 +1,3 @@
+PageTitle=Меню...
+Settings=Управление аккаунтом
+Logout=Выход
